+++ bruteforce_connect_four.c
@@ -204,7 +204,7 @@ int main( int argc, char** argv ) {
 		}
 		 
 		board* next = NULL;
-		for( int i=0; i<strlen(create_sequence); i++ ) {
+		for( size_t i=0; i<strlen(create_sequence); i++ ) {
 			int col_index = create_sequence[i] - '0';
 			printf("Drop in column %d\n", col_index);
 			next = drop( current, col_index );
