+++ utils.c
@@ -20,7 +20,7 @@ char* join( int* elements, size_t num, const char* separator ) {
 	char* buf = malloc( sizeof(char)*256 );
 	char str[100];
 	
-	for( int i=0; i<num; i++ ) {
+	for( size_t i=0; i<num; i++ ) {
 		sprintf( str, "%d", elements[i] );
 		strcat( buf, str );
 		if( i < num-1 ) {
